Licensed this work under the LGPL-3.0
My intention is for pyrisccore to be both an application and a library.
It is primarily developed as a library, a useful collection of the forms
and abstractions that make up a computer. To confirm its value as a
library, its abstractions are applied as a functional virtual machine.

Information about the LGPL-3.0:
https://www.gnu.org/licenses/lgpl-3.0.en.html

License file content copied from:
https://www.gnu.org/licenses/lgpl-3.0.txt