Lint the code with pylint
This is to start enforcing consistency as large implementation details
get worked out by the hundreds of LOC.

In my experience, the tests usually break all the rules in the interest
of readability. Rather than tediously disable messages for all or each
test, I'm opting to only run the linter on non-test source files.

Also in my experience, people appreciate a commented explaination next
to each message disabled in the config. Unless the public asks too many
questions, I'll leave the tl;dr to "personal preferences of the
maintainer" and "these are very contextual, keep an eye on them during
code review".