Added a "Slice" class
This is meant to replace Field's use of Python's "slice" object.
Python's builtin slice isn't hashable and the "step" attribute isn't
useful for defining a mask.

I'm unsure, but this might be better called "Mask". The purpose of the
object is to aid in recording the offsets of fields defined in the
RISC-V manual. The notation used (e.g. "imm[12]") is literally the slice
notation Python uses.